+++ cron/actes-menage-enveloppe.php
@@ -1,0 +1,9 @@
+#! /usr/bin/php
+<?php
+require_once(__DIR__ . "/../init/init.php");
+
+$workerScript = $objectInstancier->get(WorkerScript::class);
+$workerScript->scriptByClassName(
+    ActesMenageEnveloppeWorker::class,
+    true,
+    true);