+++ CHANGELOG.md
@@ -1,8 +1,12 @@
 # 4.3.5
 
+## Correction
+- refactoring du ménage des actes envoyés sur le cloud
+
 ## Evolution
 - mise à jour de pades-valid vers 1.4.7 #223
 - sauvegarde des mails dont le corps est vide #644
+
 
 # 4.3.4 - 2020-11-17